* Synced with the newer version of the original bitbucket memory profiler * Added a different panel to check the heap memory usage * Now you can use the tool in Unity versions < 5.6  but in those versions you can only see the heap memory usage.

+++ Assets/Editor/Treemap/Item.cs
@@ -18,7 +18,7 @@ namespace Assets.Editor.Treemap
 
         public ThingInMemory _thingInMemory;
 
-        public int memorySize { get { return _thingInMemory.size; } }
+        public long memorySize { get { return _thingInMemory.size; } }
         public string name { get { return _thingInMemory.caption; } }
         public Color color { get { return _group.color; } }
